Add support for multiple tags for multiple values
MPD supports multiple tags for some values, such as the artist, the
composer, and the performer. (See
<https://mpd.readthedocs.io/en/latest/protocol.html#tags>.) There
might be multiple artists for a track, in which case mopidy-mpd would
concatenate their names into one value, separated by a semicolon. For
some users and use cases, this can be suboptimal; when tracking the
tracks you listen to on a platform such as Listenbrainz, tracks of
multiple artists do not contribute to the total number of listens of
one of the artists.

This commit changes this behaviour to use multiple tags for values
when applicable instead of concatenating the values.